Added logic to the MemorySegmentedBufferPool out-of-band clearing.
Added logic to add the buffer being cleared back into the pending zero list if the segment can't be returned to the pool that it belongs to. This takes care of any issues of lost segments because of the segment not getting back into the pool and is the safest approach even though there is a possible, but very rare, possibility of a performance hit.